There were originally 5 Processing Completion status of which 2 were error conditions, which were NINO_NotFound and PAYE_NotFound.
Where the code set the status to NINO_NotFound and PAYE_NotFound it was changed to Completed, however, this has been removed as further logic is required to set the final completions status properly and the NotFound statuses are in an error path so just replacing them with complete does not take into account that the code is already in an error path at the point where we are trying to set the status to complete.

Story 281 will address the changes needed for error conditions.